+++ src/lib.rs
@@ -130,7 +130,7 @@
 //! ```
 
 pub mod api;
-pub mod bytes_ser_de;
+mod bytes_ser_de;
 mod cover_crypt_core;
 pub mod error;
 pub mod interfaces;